requirements.txt
Django==3.1.7

django-import-export==2.5.0

gunicorn==20.0.4
django-heroku==0.3.1

geoip2==4.1.0
